Chunk 9: split groups.js into routes/groups/ sub-routers
Refactor the 1,464-line routes/groups.js into a sub-router directory
following the backup/members/portal precedent. Pure move, no behaviour
change:

- list.js     — group listing + list-level Excel/PDF download
- crud.js     — single-group fetch/create/update/delete
- members.js  — membership listing, download, add/remove and bulk ops
- events.js   — group schedule sub-resource
- ledger.js   — group-ledger sub-resource (incl. hasLedgerAccess)
- index.js    — shared requireAuth + requireFeature('groups'); mounts the
                list router before the /:id CRUD router to preserve Express
                match order

app.js now imports groups/index.js. Backend suite green (593 tests),
lint and Prettier clean.
